Create gyro turn testing code
Team WT used this to test their gyros in their brand new Spike Prime bots in late 2023. We noticed some really odd turns going on with the gyro, including some random spins on the table. This code helped the team determine that we should NOT use the gyro, after all. It worked better to use a correction factor and no gyro. (but no, still not perfect). WT would prefer to use a gyro that worked!